Move test to front of run, provide auth and anon versions
